Improves error handling of script entrypoint
The returns of async functions *should* always be taken care of.
Simply using the `#catch`-callback to log the error
and printing a generic error message to the user should be enough.

Maybe we can replace the alert-call with showing a error-box on the page.
But it is probably only visible in very outdated or weird browsers